fix type annotations for FullBlock.header_hash and FullBlock.prev_header_hash (#10909)

+++ chia/consensus/block_body_validation.py
@@ -346,7 +346,7 @@ async def validate_block_body(
                 )
             if curr.height == 0:
                 break
-            curr = reorg_blocks[curr.height - 1]
+            curr = reorg_blocks[uint32(curr.height - 1)]
             assert curr is not None
 
     removal_coin_records: Dict[bytes32, CoinRecord] = {}
